requirements.txt
backtrader
pandas
matplotlib
google-genai
